Add test for `migrate_cache_entries_to_double_dashes`.
